requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
opencv-contrib-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
onnxruntime==1.19.2
pgvector==0.3.3
mediapipe==0.10.14